+++ models/feeProject/PeriodReceptFee.js
@@ -17,15 +17,15 @@ module.exports = (sequelize, DataTypes) => (
             allowNull: true,
         },
         basicCost: { // 기본 요금
-            type: DataTypes.INTEGER(11),
+            type: DataTypes.BIGINT,
             allowNull: true,
         },
         kwCost: { // Kw 당 요금
-            type: DataTypes.INTEGER(11),
+            type: DataTypes.BIGINT,
             allowNull: true, 
         },
         sum: { // 소계
-            type: DataTypes.INTEGER(11),
+            type: DataTypes.BIGINT,
             allowNull: true,
         },
         createdAt: { // 생성 시간
