+++ models/product/BucketProductOpt.js
@@ -5,7 +5,7 @@ module.exports = (sequelize, DataTypes) => (
             allowNull: false,
         },
         price: { // 가격
-            type: DataTypes.INTEGER(11),
+            type: DataTypes.BIGINT,
             allowNull: false,
         },
         num: { // 수량
@@ -13,7 +13,7 @@ module.exports = (sequelize, DataTypes) => (
             allowNull: false,
         },
         totalPrice: { // 총 가격
-            type: DataTypes.INTEGER(11),
+            type: DataTypes.BIGINT,
             allowNull: true,
         },
     }, {
